Reference

API

Variables

MultivariatePolynomials.name — Function.

name(v::AbstractVariable)::AbstractString

Returns the name of a variable.

source

Terms

term
zeroterm
coefficient
monomial
exponent
deg
isconstant
divides

Polynomials

terms
monomials
mindeg
maxdeg
extdeg
leadingterm
leadingcoefficient
leadingmonomial
removeleadingterm
removemonomials
vars
nvars

Monomial Vectors

MultivariatePolynomials.monovec — Function.

monovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the vector of monomials X in decreasing order and without any duplicates.

Examples

Calling monovec on $[xy, x, xy, x^2y, x]$ should return $[x^2y, xy, x]$.

source

MultivariatePolynomials.monovectype — Function.

monovectype(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the return type of monovec.

source

MultivariatePolynomials.sortmonovec — Function.

sortmonovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns σ, the orders in which one must take the monomials in X to make them sorted and without any duplicate and the sorted vector of monomials, i.e. it returns (σ, X[σ]).

Examples

Calling sortmonovec on $[xy, x, xy, x^2y, x]$ should return $([4, 1, 2], [x^2y, xy, x])$.

source

MultivariatePolynomials.mergemonovec — Function.

mergemonovec{MT<:AbstractMonomialLike, MVT<:AbstractVector{MT}}(X::AbstractVector{MVT}}

Returns the vector of monomials in the entries of X in decreasing order and without any duplicates, i.e. monovec(vcat(X...))

Examples

Calling mergemonovec on $[[xy, x, xy], [x^2y, x]]$ should return $[x^2y, xy, x]$.

source