Reference

API

Variables

MultivariatePolynomials.name — Function.

name(v::AbstractVariable)::AbstractString

Returns the name of a variable.

source

Monomials

MultivariatePolynomials.nvariables — Function.

nvariables(p::AbstractPolynomialLike)

Returns the number of variables in p, i.e. length(variables(p)). It could be more than the number of variables with nonzero exponent (see the Examples section of variables).

Examples

Calling nvariables(x^2*y) should return at least 2 and calling nvariables(x) should return at least 1.

source

MultivariatePolynomials.variables — Function.

variables(p::AbstractPolynomialLike)

Returns the tuple of the variables of p in decreasing order. It could contain variables of zero degree, see the example section.

Examples

Calling variables(x^2*y) should return (x, y) and calling variables(x) should return (x,). Note that the variables of m does not necessarily have nonzero exponent. For instance, variables([x^2*y, y*z][1]) is usually (x, y, z) since the two monomials have been promoted to a common type.

source

MultivariatePolynomials.exponents — Function.

exponents(t::AbstractTermLike)

Returns the exponent of the variables in the monomial of the term t.

Examples

Calling exponents(x^2*y) should return (2, 1).

source

MultivariatePolynomials.exponent — Function.

exponent(t::AbstractTermLike, var::AbstractVariable)

Returns the exponent of the variable var in the monomial of the term t.

Examples

Calling exponent(x^2*y, x) should return 2 and calling exponent(x^2*y, y) should return 1.

source

MultivariatePolynomials.deg — Function.

deg(t::AbstractTermLike)

Returns the total degree of the monomial of the term t, i.e. sum(exponents(t)).

Examples

Calling deg(x^2*y) should return 3 which is $2 + 1$.

source

MultivariatePolynomials.isconstant — Function.

isconstant(m::AbstractMonomialLike)

Returns whether the monomial is constant.

source

MultivariatePolynomials.constantmonomial — Function.

constantmonomial(p::AbstractPolynomialType)

Returns a constant monomial of the monomial type of p with the same variables as p.

constantmonomial(::Type{PT}) where {PT<:AbstractPolynomialType}

Returns a constant monomial of the monomial type of a polynomial of type PT.

source

MultivariatePolynomials.divides — Function.

divides(t1::AbstractTermLike, t2::AbstractTermLike)

Returns whether monomial(t1) divides monomial(t2).

Examples

Calling divides(2x^2y, 3xy) should return false because x^2y does not divide xy since x has a degree 2 in x^2y which is greater than the degree of x on xy. However, calling divides(3xy, 2x^2y) should return true.

source

Terms

MultivariatePolynomials.term — Function.

term(p::AbstractPolynomialLike)

Converts the polynomial p to a term. When applied on a polynomial, it throws an error if it has more than one term. When applied to a term, it is the identity and does not copy it. When applied to a monomial, it create a term of type AbstractTerm{Int}.

source

MultivariatePolynomials.coefficient — Function.

coefficient(t::AbstractTermLike)

Returns the coefficient of the term t.

Examples

Calling coefficient on $4x^2y$ should return $4$.

source

MultivariatePolynomials.monomial — Function.

monomial(t::AbstractTermLike)

Returns the monomial of the term t.

Examples

Calling monomial on $4x^2y$ should return $x^2y$.

source

Polynomials

MultivariatePolynomials.terms — Function.

terms(p::AbstractPolynomialLike)

Returns an iterator over the nonzero terms of the polynomial p sorted in the decreasing monomial order.

Examples

Calling terms on $4x^2y + xy + 2x$ should return an iterator of $[4x^2y, xy, 2x]$.

source

MultivariatePolynomials.monomials — Function.

monomials(p::AbstractPolynomialLike)

Returns an iterator over the monomials of p of the nonzero terms of the polynomial sorted in the decreasing order.

monomials(vars::Tuple, degs::AbstractVector{Int}, filter::Function = m -> true)

Builds the vector of all the monovec m with variables vars such that the degree deg(m) is in degs and filter(m) is true.

Examples

Calling monomials on $4x^2y + xy + 2x$ should return an iterator of $[x^2y, xy, x]$.

Calling monomials((x, y), [1, 3], m -> exponent(m, y) != 1) should return [x^3, x*y^2, y^3, x] where x^2*y and y have been excluded by the filter.

source

MultivariatePolynomials.mindeg — Function.

mindeg(p::AbstractPolynomialLike)

Returns the minimal total degree of the monomials of p, i.e. minimum(deg, terms(p)).

Examples

Calling mindeg on on $4x^2y + xy + 2x$ should return 1.

source

MultivariatePolynomials.maxdeg — Function.

maxdeg(p::AbstractPolynomialLike)

Returns the maximal total degree of the monomials of p, i.e. maximum(deg, terms(p)).

Examples

Calling maxdeg on on $4x^2y + xy + 2x$ should return 3.

source

MultivariatePolynomials.extdeg — Function.

extdeg(p::AbstractPolynomialLike)

Returns the extremal total degrees of the monomials of p, i.e. (mindeg(p), maxdeg(p)).

Examples

Calling extdeg on on $4x^2y + xy + 2x$ should return (1, 3).

source

MultivariatePolynomials.leadingterm — Function.

leadingterm(p::AbstractPolynomialLike)

Returns the coefficient of the leading term, i.e. first(terms(p)).

Examples

Calling leadingterm on $4x^2y + xy + 2x$ should return $4x^2y$.

source

MultivariatePolynomials.leadingcoefficient — Function.

leadingcoefficient(p::AbstractPolynomialLike)

Returns the coefficient of the leading term of p, i.e. coefficient(leadingterm(p)).

Examples

Calling leadingcoefficient on $4x^2y + xy + 2x$ should return $4$ and calling it on $0$ should return $0$.

source

MultivariatePolynomials.leadingmonomial — Function.

leadingmonomial(p::AbstractPolynomialLike)

Returns the monomial of the leading term of p, i.e. monomial(leadingterm(p)) or first(monomials(p)).

Examples

Calling leadingmonomial on $4x^2y + xy + 2x$ should return $x^2y$.

source

MultivariatePolynomials.removeleadingterm — Function.

removeleadingterm(p::AbstractPolynomialLike)

Returns a polynomial with the leading term removed in the polynomial p.

Examples

Calling removeleadingterm on $4x^2y + xy + 2x$ should return $xy + 2x$.

source

MultivariatePolynomials.removemonomials — Function.

Returns a polynomial with the terms having their monomial in the monomial vector mv removed in the polynomial p.

Examples

Calling removemonomials(4x^2*y + x*y + 2x, [x*y]) should return $4x^2*y + 2x$.

source

Monomial Vectors

MultivariatePolynomials.monovec — Function.

monovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the vector of monomials X in decreasing order and without any duplicates.

Examples

Calling monovec on $[xy, x, xy, x^2y, x]$ should return $[x^2y, xy, x]$.

source

MultivariatePolynomials.monovectype — Function.

monovectype(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the return type of monovec.

source

MultivariatePolynomials.sortmonovec — Function.

sortmonovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns σ, the orders in which one must take the monomials in X to make them sorted and without any duplicate and the sorted vector of monomials, i.e. it returns (σ, X[σ]).

Examples

Calling sortmonovec on $[xy, x, xy, x^2y, x]$ should return $([4, 1, 2], [x^2y, xy, x])$.

source

MultivariatePolynomials.mergemonovec — Function.

mergemonovec{MT<:AbstractMonomialLike, MVT<:AbstractVector{MT}}(X::AbstractVector{MVT}}

Returns the vector of monomials in the entries of X in decreasing order and without any duplicates, i.e. monovec(vcat(X...))

Examples

Calling mergemonovec on $[[xy, x, xy], [x^2y, x]]$ should return $[x^2y, xy, x]$.

source