Reference

API

Variables

MultivariatePolynomials.name — Function.

name(v::AbstractVariable)::AbstractString

Returns the name of a variable.

source

MultivariatePolynomials.similarvariable — Function.

similarvariable(p::AbstractPolynomialLike, variable::Type{Val{V}})

Creates a new variable V based upon the the given source polynomial.

similarvariable(p::AbstractPolynomialLike, v::Symbol)

Creates a new variable based upon the given source polynomial and the given symbol v. Note that this can lead to type instabilities.

Examples

Calling similarvariable(typedpoly, Val{:x}) on a polynomial created with TypedPolynomials results in TypedPolynomials.Variable{:x}.

source

Monomials

nvariables
variables
exponents
exponent
deg
isconstant
constantmonomial
divides

Terms

MultivariatePolynomials.term — Function.

term(p::AbstractPolynomialLike)

Converts the polynomial p to a term. When applied on a polynomial, it throws an error if it has more than one term. When applied to a term, it is the identity and does not copy it. When applied to a monomial, it create a term of type AbstractTerm{Int}.

source

MultivariatePolynomials.coefficient — Function.

coefficient(t::AbstractTermLike)

Returns the coefficient of the term t.

Examples

Calling coefficient on $4x^2y$ should return $4$.

source

MultivariatePolynomials.monomial — Function.

monomial(t::AbstractTermLike)

Returns the monomial of the term t.

Examples

Calling monomial on $4x^2y$ should return $x^2y$.

source

Polynomials

terms
monomials
mindeg
maxdeg
extdeg
leadingterm
leadingcoefficient
leadingmonomial
removeleadingterm
removemonomials

Monomial Vectors

MultivariatePolynomials.monovec — Function.

monovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the vector of monomials X in decreasing order and without any duplicates.

Examples

Calling monovec on $[xy, x, xy, x^2y, x]$ should return $[x^2y, xy, x]$.

source

MultivariatePolynomials.monovectype — Function.

monovectype(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns the return type of monovec.

source

MultivariatePolynomials.sortmonovec — Function.

sortmonovec(X::AbstractVector{MT}) where {MT<:AbstractMonomialLike}

Returns σ, the orders in which one must take the monomials in X to make them sorted and without any duplicate and the sorted vector of monomials, i.e. it returns (σ, X[σ]).

Examples

Calling sortmonovec on $[xy, x, xy, x^2y, x]$ should return $([4, 1, 2], [x^2y, xy, x])$.

source

MultivariatePolynomials.mergemonovec — Function.

mergemonovec{MT<:AbstractMonomialLike, MVT<:AbstractVector{MT}}(X::AbstractVector{MVT}}

Returns the vector of monomials in the entries of X in decreasing order and without any duplicates, i.e. monovec(vcat(X...))

Examples

Calling mergemonovec on $[[xy, x, xy], [x^2y, x]]$ should return $[x^2y, xy, x]$.

source