Internal

Internal utilities:

MultivariatePolynomials.pair_zip — Function

pair_zip((a, b), (c, d)) gives (a=>c, b=>d)

This function was written by Fengyang Wang and shared on the Julia discourse forum: https://discourse.julialang.org/t/type-stable-zip-to-pairs/3390/2

Base.vec — Function

Convert a tuple of variables into a static vector to allow array-like usage. The element type of the vector will be Monomial{vars, length(vars)}.

MultivariatePolynomials.print_maybe_multiplication_sign — Function

print_maybe_multiplication_sign(io, mime)

Prints a multiplication sign depending on the mime type.

MultivariatePolynomials.LazyMap — Type

struct LazyMap{T, VT}
    f::Function
    data::VT
end

Iterator over the elements of data mapped by f. This is similar to Base.Generator(f, data) except that the eltype of a LazyMap is given at construction while the eltype of Base.Generator(f, data) is Any.